+++ src/BuildingTypes/Building.java
@@ -28,5 +28,5 @@ public abstract class Building {
 	 * This method is for updating the building type
 	 * @return Whether or not the building could be successfully upgraded
 	 */
-	public abstract double upgrade(double money);
+	public abstract int upgrade(double money);
 }
