+++ src/BuildingTypes/Commerce.java
@@ -38,18 +38,18 @@ public class Commerce extends Building {
 
 
 	@Override
-	public double upgrade(double money) {
+	public int upgrade(double money) {
 		if (level < 3) {
 			switch (level){
 			case 1:
-				if (money >= 250.0) {
+				if (money >= 250) {
 					level++;
-					return 250.0;
+					return 250;
 				}
 				break;
 			}
 		}
-		return -1.0;
+		return -1;
 	}
 	
 	@Override
