+++ src/BuildingTypes/Defense.java
@@ -37,18 +37,18 @@ public class Defense extends Building {
 
 
 	@Override
-	public double upgrade(double money) {
+	public int upgrade(double money) {
 		if (level < 3) {
 			switch (level){
 			case 1:
-				if (money >= 500.0) {
+				if (money >= 500) {
 					level++;
-					return 500.0;
+					return 500;
 				}
 				break;
 			}
 		}
-		return -1.0;
+		return -1;
 	}
 	
 	@Override
