+++ src/BuildingTypes/Shelter.java
@@ -41,24 +41,24 @@ public class Shelter extends Building {
 
 
 	@Override
-	public double upgrade(double money) {
+	public int upgrade(double money) {
 		if (level < 4) {
 			switch (level){
 			case 1:
 				if (money >= 100.0) {
 					level++;
-					return 100.0;
+					return 100;
 				}
 				break;
 			case 2: 
 				if (money >= 500.0) {
 					level++;
-					return 500.0;
+					return 500;
 				}
 				break;
 			}
 		}
-		return -1.0;
+		return -1;
 	}
 	
 	@Override
